refactor: move FloatingBox to 'components/containers'
