+++ src/components/containers/FloatingBox.tsx
