+++ src/pages/unstake.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next';
-import { FloatingBox } from 'src/layout/FloatingBox';
+import { FloatingBox } from 'src/components/containers/FloatingBox';
 
 const Unstake: NextPage = () => {
   return <FloatingBox>Unstake!</FloatingBox>;
